Inference lab Assigment
This is Document was design as a part of interview of Inference lab 2nd Assigment.

Q : Design an approach for developing customized promotional offers for customers of a particular bank (say ICICI). The promotional offers must be curated/tailored for each user.
Please email the approach with screenshots of a possible UI as well (hand-drawn) is acceptable.